pyproject.toml
[tool.poetry]
name = "api_rest"
version = "0.1.0"
description = ""
authors = ["Everton Arakaki <everton.arakaki@shipay.com.br>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"